+++ Changelog.md
@@ -6,7 +6,8 @@
 
 * Support for new syntax of PostgreSQL 19 (as of beta 3):
   * `graph_table()` construct in the `FROM` clause allows querying newly added property graphs;
-  * `INSERT ... ON CONFLICT ... DO SELECT`.
+  * `INSERT ... ON CONFLICT ... DO SELECT`;
+  * `IGNORE NULLS` / `RESPECT NULLS` setting for window functions.
 * Tested on PHP 8.6
 
 ### Changed